changes to final version
