+++ src/components/analysisCharts/earnings_chart/EarningsChart.jsx
@@ -34,7 +34,7 @@ function AnalysisPage({ data }) {
     ];
 
     return (
-        <ResponsiveContainer height={400} className="mt-5">
+        <ResponsiveContainer height={400} className="mt-5" style={{ marginTop: '25px' }}>
             <h3 style={{ textAlign: 'center' }}>Analiza zarobków</h3>
             <div style={{ display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
                 <div style={{ width: '50%' }}>
