+++ src/pages/home_page/home_page.css
@@ -1,0 +1,112 @@
+.main_page {
+    width: 100%;
+    height: 100vh;
+    background-color: #f4f7f6;
+    font-family: 'Open Sans', sans-serif;
+}
+
+.homepage-header {
+    display: flex;
+    flex-direction: row; /* Kontynuujemy użycie flexboxa w poziomie */
+    align-items: start; /* Wyrownujemy do góry */
+    padding: 2rem;
+    background: linear-gradient(to right, #e3ffe1, #effff8);
+    flex-wrap: wrap; /* Pozwól na zawijanie elementów jeśli nie mieszczą się w jednym rzędzie */
+    height:70vh;
+}
+
+
+.header-content {
+    width: 100%; /* Zajmij całą szerokość w nowym rzędzie */
+    text-align: center; /* Wyśrodkuj tekst */
+    margin-bottom: 1rem; /* Dodaj trochę miejsca na dole */
+}
+
+.job-search {
+    display: flex; /* Stosujemy flexbox dla lepszego kontrolowania elementów formularza */
+    flex: 1; /* Formularz będzie się rozciągać, ale nie więcej niż obrazek */
+    align-items: center; /* Elementy wewnątrz będą wyśrodkowane wertykalnie */
+}
+
+.header-image {
+    flex: 1; /* Obrazek będzie się rozciągać, ale nie więcej niż formularz */
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+/* Upewnijmy się, że poszczególne elementy formularza są dobrze wyświetlane obok siebie */
+.form-inline {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+
+.job-search .form-control,
+.job-search .form-select,
+.job-search .home-button {
+    margin: 0 0.5rem; /* Dodaj trochę przestrzeni między elementami */
+}
+
+  
+  .header-image {
+    background-image: url('../../assets/home_page.png');
+    background-size: contain;
+    background-repeat: no-repeat;
+    background-position: center;
+    width: 30%; 
+    height: 100%; 
+    border-radius: 0; 
+    box-shadow: none; 
+  }
+  
+  .homepage-header h1 {
+    font-size: 2.5rem;
+    color: #333; 
+    margin-bottom: 0.5rem;
+  }
+  
+  .homepage-header p {
+    font-size: 1.25rem;
+    color: #555; 
+    margin-bottom: 1rem;
+  }
+  
+  .search-bar-container {
+    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2); 
+    border-radius: 6px; 
+    height:100px;
+  }
+
+  .form-control, .form-select {
+    border: 1px solid #ced4da; 
+    box-shadow: 0 2px 4px 0 rgba(0,0,0,0.1); 
+    transition: box-shadow 0.3s ease-in-out; 
+  }
+  
+  .form-control:focus, .form-select:focus {
+    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2); 
+  }
+  
+  .home-button {
+    background-color: #7b2ff7; 
+    color: white; 
+    border: none; 
+    padding: 0.5rem 1rem; 
+    margin-left: 1.3rem;
+    border-radius: 4px; 
+    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.15);
+  }
+  
+  @media (max-width: 768px) {
+    .homepage-header {
+        flex-direction: column;
+    }
+  
+    .header-content,
+    .job-search,
+    .header-image {
+        width: 100%; /* Na urządzeniach mobilnych każdy element zajmuje 100% szerokości */
+        margin-bottom: 1rem; /* Dodajmy trochę miejsca między elementami */
+    }
+}
