+++ src/services/api/jobOffersService.jsx
@@ -10,7 +10,8 @@ const handleResponse = (response) => {
 
 const fetchAllJobOffers = async (page) => {
     try {
-        const response = await axios.get(`https://jobbufferback.azurewebsites.net/api/offers/`, { params: { page } });
+        // const response = await axios.get(`https://jobbufferback.azurewebsites.net/api/offers/`, { params: { page } });
+        const response = await axios.get(`http://127.0.0.1:8000/api/offers/`, { params: { page } });
         return handleResponse(response);
     } catch (error) {
         throw new Error(error.response?.data?.message || "Wystąpił błąd podczas pobierania wszystkich ofert pracy.");
@@ -19,7 +20,8 @@ const fetchAllJobOffers = async (page) => {
 
 const fetchCategories = async () => {
     try {
-        const response = await axios.get('https://jobbufferback.azurewebsites.net/api/categories/');
+        // const response = await axios.get('https://jobbufferback.azurewebsites.net/api/categories/');
+        const response = await axios.get('http://127.0.0.1:8000/api/categories/');
         console.log(response);
         return response.data;
     } catch (error) {
@@ -42,7 +44,8 @@ const filterAllJobOffers = async (categoryName, query, jobLocation, province, cu
             params.set('jobLocation', jobLocation);
         }
         params.set('page', currentPage.toString());
-        const response = await axios.get(`http://localhost:8000/api/offers/filter/`, { params });
+        // const response = await axios.get(`https://jobbufferback.azurewebsites.net/api/offers/filter/`, { params });
+        const response = await axios.get(`http://127.0.0.1:8000/api/offers/filter/`, { params });
         return handleResponse(response);
     } catch (error) {
         console.error('Błąd podczas wykonywania żądania do API:', error);
@@ -52,7 +55,8 @@ const filterAllJobOffers = async (categoryName, query, jobLocation, province, cu
 
 const analysisByJobOffersWebiste = async () => {
     try {
-        const response = await axios.get('http://localhost:8000/api/analysis/job_offers');
+        // const response = await axios.get('https://jobbufferback.azurewebsites.net/api/analysis/job_offers');
+        const response = await axios.get('http://127.0.0.1:8000/api/analysis/job_offers');
         return response.data;
     } catch (error) {
         console.error('Error fetching job offers analysis data: ', error);
@@ -62,7 +66,8 @@ const analysisByJobOffersWebiste = async () => {
 
 const analysisByJobOffersCategory = async () => {
     try {
-        const response = await axios.get('http://localhost:8000/api/analysis/job_offers_category/');
+        // const response = await axios.get('https://jobbufferback.azurewebsites.net/api/analysis/job_offers_category/');
+        const response = await axios.get('http://127.0.0.1:8000/api/analysis/job_offers_category/');
         return response.data;
     } catch (error) {
         console.error('Error fetching job offers category analysis data: ', error);
@@ -72,7 +77,8 @@ const analysisByJobOffersCategory = async () => {
 
 const analysisByAverageEarnings = async () => {
     try {
-        const response = await axios.get('http://localhost:8000/api/analysis/average_earnings/');
+        // const response = await axios.get('https://jobbufferback.azurewebsites.net/api/analysis/average_earnings/');
+        const response = await axios.get('http://127.0.0.1:8000/api/analysis/average_earnings/');
         return (response.data);
     } catch (error) {
         console.error('Error fetching earnings data:', error);
@@ -82,7 +88,8 @@ const analysisByAverageEarnings = async () => {
 
 const analysisByEarningsHeatmap = async () => {
     try {
-        const response = await axios.get('http://localhost:8000/api/analysis/earnings_heatmap/');
+        // const response = await axios.get('https://jobbufferback.azurewebsites.net/api/analysis/earnings_heatmap/');
+        const response = await axios.get('http://127.0.0.1:8000/api/analysis/earnings_heatmap/');
         return response.data;
     } catch (error) {
         console.error('Error fetching earnings heatmap data: ', error);
@@ -92,7 +99,8 @@ const analysisByEarningsHeatmap = async () => {
 
 const analysisByJobOffersPerDay = async () => {
     try {
-        const response = await axios.get('http://localhost:8000/api/analysis/job_offers_by_day/');
+        // const response = await axios.get('https://jobbufferback.azurewebsites.net/api/analysis/job_offers_by_day/');
+        const response = await axios.get('http://127.0.0.1:8000/api/analysis/job_offers_by_day/');
         return response.data;
     } catch (error) {
         console.error('Error fetching data: ', error);
